Clarify which bit will be added
Most sections described in this post are already well outlined in the
manual.

Unfortunately, I'm not really sure where to start. One unfortunate
blister with Nix is that the wiki is currently shut down, and this:
https://github.com/NixOS/nixpkgs/issues/22599

is still an issue as of this commit. It probably would be better in a
wiki system anyway.